requirements.txt
steam[client] @ git+https://github.com/detiam/steam_websocket@b8239912e6a190f490aede529c08b5049096bdc8
pyinstaller
requests
certifi
python-dotenv
PyYAML
httpx
tqdm
pycryptodome
rich
tenacity
